[BE][DTensor] move torch.distributed._tensor import to torch.distributed.tensor in test files
[ghstack-poisoned]
