+++ test/distributed/_composable/fsdp/test_fully_shard_overlap.py
@@ -7,8 +7,8 @@ from typing import Callable
 import torch
 import torch.distributed as dist
 import torch.nn as nn
-from torch.distributed._tensor.experimental import implicit_replication
 from torch.distributed.fsdp import fully_shard
+from torch.distributed.tensor.experimental import implicit_replication
 from torch.testing._internal.common_distributed import skip_if_lt_x_gpu
 from torch.testing._internal.common_fsdp import (
     FSDPTest,
