+++ test/distributed/tensor/experimental/test_tp_transform.py
@@ -2,7 +2,7 @@
 from collections import defaultdict
 
 import torch
-from torch.distributed._tensor.experimental._tp_transform import (
+from torch.distributed.tensor.experimental._tp_transform import (
     tensor_parallel_transformation,
 )
 from torch.distributed.tensor.parallel.style import (
