+++ test/distributed/tensor/test_api.py
@@ -3,7 +3,7 @@
 
 import torch
 import torch.nn as nn
-from torch.distributed._tensor import (
+from torch.distributed.tensor import (
     DeviceMesh,
     distribute_module,
     distribute_tensor,
