+++ test/distributed/tensor/test_xla_integration.py
@@ -10,7 +10,7 @@ import numpy as np
 
 import torch
 from torch import nn
-from torch.distributed._tensor import (
+from torch.distributed.tensor import (
     DeviceMesh,
     distribute_module,
     distribute_tensor,
